+++ Moco/Moco/MocoProblem.h
@@ -382,7 +382,7 @@ protected: // Protected so that doxygen shows the properties.
             "Parameter variables (model properties) to optimize.");
     OpenSim_DECLARE_LIST_PROPERTY(
             goals, MocoGoal,
-            "Integral/endpoint quantities to minimize or constrain.");
+            "Integral/boundary quantities to minimize or constrain.");
     OpenSim_DECLARE_LIST_PROPERTY(path_constraints, MocoPathConstraint,
             "Path constraints to enforce in the optimal control problem.");
     // TODO make this a list property of MocoConstraintInfos when we are able to
@@ -411,7 +411,7 @@ private:
 ///   - OpenSim Model
 ///   - state and control variable info (e.g., bounds)
 ///   - parameter variables (model properties)
-///   - goals (costs and endpoint constraints)
+///   - goals (costs and boundary constraints)
 ///   - path constraints
 ///
 /// Currently, only single-phase problems are supported.