+++ CHANGELOG.md
@@ -1,3 +1,5 @@
+- 2019-11-03: Endpoint constraints were renamed to boundary constraints.
+
 - 2019-10-19: configureMoco.m adds Moco's Matlab Utilities directory
               to the Matlab path, and removes any detected OpenSense beta
               installations from Matlab.